Normalize OpenTelemetry annotation and instrumentation API javaagent packages
