--- instrumentation/opentelemetry-extension-annotations-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/extensionannotations/v1_0/WithSpanSingletons.java
@@ -1,100 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.extensionannotations.v1_0;
-
-import static java.util.logging.Level.FINE;
-
-import io.opentelemetry.api.GlobalOpenTelemetry;
-import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.instrumentation.api.annotation.support.MethodSpanAttributesExtractor;
-import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesExtractor;
-import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.api.semconv.util.SpanNames;
-import java.lang.reflect.Method;
-import java.util.logging.Logger;
-
-@SuppressWarnings("deprecation") // instrumenting deprecated class for backwards compatibility
-public class WithSpanSingletons {
-
-  private static final String INSTRUMENTATION_NAME =
-      "io.opentelemetry.opentelemetry-extension-annotations-1.0";
-
-  private static final Logger logger = Logger.getLogger(WithSpanSingletons.class.getName());
-  private static final Instrumenter<Method, Object> instrumenter = createInstrumenter();
-  private static final Instrumenter<MethodRequest, Object> instrumenterWithAttributes =
-      createInstrumenterWithAttributes();
-
-  public static Instrumenter<Method, Object> instrumenter() {
-    return instrumenter;
-  }
-
-  public static Instrumenter<MethodRequest, Object> instrumenterWithAttributes() {
-    return instrumenterWithAttributes;
-  }
-
-  private static Instrumenter<Method, Object> createInstrumenter() {
-    return Instrumenter.builder(
-            GlobalOpenTelemetry.get(), INSTRUMENTATION_NAME, WithSpanSingletons::spanNameFromMethod)
-        .addAttributesExtractor(CodeAttributesExtractor.create(new MethodCodeAttributesGetter()))
-        .buildInstrumenter(WithSpanSingletons::spanKindFromMethod);
-  }
-
-  private static Instrumenter<MethodRequest, Object> createInstrumenterWithAttributes() {
-    return Instrumenter.builder(
-            GlobalOpenTelemetry.get(),
-            INSTRUMENTATION_NAME,
-            WithSpanSingletons::spanNameFromMethodRequest)
-        .addAttributesExtractor(
-            CodeAttributesExtractor.create(new MethodRequestCodeAttributesGetter()))
-        .addAttributesExtractor(
-            MethodSpanAttributesExtractor.create(
-                MethodRequest::method,
-                new WithSpanParameterAttributeNamesExtractor(),
-                MethodRequest::args))
-        .buildInstrumenter(WithSpanSingletons::spanKindFromMethodRequest);
-  }
-
-  private static SpanKind spanKindFromMethodRequest(MethodRequest request) {
-    return spanKindFromMethod(request.method());
-  }
-
-  private static SpanKind spanKindFromMethod(Method method) {
-    application.io.opentelemetry.extension.annotations.WithSpan annotation =
-        method.getDeclaredAnnotation(
-            application.io.opentelemetry.extension.annotations.WithSpan.class);
-    if (annotation == null) {
-      return SpanKind.INTERNAL;
-    }
-    return toAgentOrNull(annotation.kind());
-  }
-
-  private static SpanKind toAgentOrNull(
-      application.io.opentelemetry.api.trace.SpanKind applicationSpanKind) {
-    try {
-      return SpanKind.valueOf(applicationSpanKind.name());
-    } catch (IllegalArgumentException e) {
-      logger.log(FINE, "unexpected span kind: {0}", applicationSpanKind.name());
-      return SpanKind.INTERNAL;
-    }
-  }
-
-  private static String spanNameFromMethodRequest(MethodRequest request) {
-    return spanNameFromMethod(request.method());
-  }
-
-  private static String spanNameFromMethod(Method method) {
-    application.io.opentelemetry.extension.annotations.WithSpan annotation =
-        method.getDeclaredAnnotation(
-            application.io.opentelemetry.extension.annotations.WithSpan.class);
-    String spanName = annotation.value();
-    if (spanName.isEmpty()) {
-      spanName = SpanNames.fromMethod(method);
-    }
-    return spanName;
-  }
-
-  private WithSpanSingletons() {}
-}
