+++ instrumentation/opentelemetry-extension-annotations-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/opentelemetry/extension/annotations/v1_0/MethodRequest.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.opentelemetry.extension.annotations.v1_0;
+
+import java.lang.reflect.Method;
+
+class MethodRequest {
+  private final Method method;
+  private final Object[] args;
+
+  MethodRequest(Method method, Object[] args) {
+    this.method = method;
+    this.args = args;
+  }
+
+  Method method() {
+    return this.method;
+  }
+
+  Object[] args() {
+    return this.args;
+  }
+}
