+++ instrumentation/opentelemetry-extension-annotations-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/opentelemetry/extension/annotations/v1_0/MethodRequestCodeAttributesGetter.java
@@ -1,0 +1,21 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.opentelemetry.extension.annotations.v1_0;
+
+import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesGetter;
+
+final class MethodRequestCodeAttributesGetter implements CodeAttributesGetter<MethodRequest> {
+
+  @Override
+  public Class<?> getCodeClass(MethodRequest methodRequest) {
+    return methodRequest.method().getDeclaringClass();
+  }
+
+  @Override
+  public String getMethodName(MethodRequest methodRequest) {
+    return methodRequest.method().getName();
+  }
+}
