--- instrumentation/opentelemetry-instrumentation-annotations-1.16/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/instrumentationannotations/v1_16/AnnotationSingletons.java
@@ -1,156 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.instrumentationannotations.v1_16;
-
-import static java.util.logging.Level.FINE;
-
-import io.opentelemetry.api.GlobalOpenTelemetry;
-import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.context.Context;
-import io.opentelemetry.instrumentation.api.annotation.support.MethodSpanAttributesExtractor;
-import io.opentelemetry.instrumentation.api.annotation.support.SpanAttributesExtractor;
-import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesExtractor;
-import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.api.semconv.util.SpanNames;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.MethodType;
-import java.lang.reflect.Method;
-import java.util.logging.Logger;
-import javax.annotation.Nullable;
-
-public class AnnotationSingletons {
-
-  private static final String INSTRUMENTATION_NAME =
-      "io.opentelemetry.opentelemetry-instrumentation-annotations-1.16";
-
-  private static final Logger logger = Logger.getLogger(AnnotationSingletons.class.getName());
-  private static final Instrumenter<Method, Object> instrumenter = createInstrumenter();
-  private static final Instrumenter<MethodRequest, Object> instrumenterWithAttributes =
-      createInstrumenterWithAttributes();
-  private static final SpanAttributesExtractor attributes = createAttributesExtractor();
-
-  // The reason for using reflection here is that it needs to be compatible with the old version of
-  // @WithSpan annotation that does not include the inheritContext option to avoid failing the
-  // muzzle check.
-  @Nullable
-  private static final MethodHandle INHERIT_CONTEXT_METHOD_HANDLE =
-      findInheritContextMethodHandle();
-
-  @Nullable
-  private static MethodHandle findInheritContextMethodHandle() {
-    try {
-      return MethodHandles.publicLookup()
-          .findVirtual(
-              application.io.opentelemetry.instrumentation.annotations.WithSpan.class,
-              "inheritContext",
-              MethodType.methodType(boolean.class));
-    } catch (NoSuchMethodException | IllegalAccessException ignored) {
-      return null;
-    }
-  }
-
-  public static Instrumenter<Method, Object> instrumenter() {
-    return instrumenter;
-  }
-
-  public static Instrumenter<MethodRequest, Object> instrumenterWithAttributes() {
-    return instrumenterWithAttributes;
-  }
-
-  public static SpanAttributesExtractor attributes() {
-    return attributes;
-  }
-
-  private static Instrumenter<Method, Object> createInstrumenter() {
-    return Instrumenter.builder(
-            GlobalOpenTelemetry.get(),
-            INSTRUMENTATION_NAME,
-            AnnotationSingletons::spanNameFromMethod)
-        .addAttributesExtractor(CodeAttributesExtractor.create(new MethodCodeAttributesGetter()))
-        .buildInstrumenter(AnnotationSingletons::spanKindFromMethod);
-  }
-
-  private static Instrumenter<MethodRequest, Object> createInstrumenterWithAttributes() {
-    return Instrumenter.builder(
-            GlobalOpenTelemetry.get(),
-            INSTRUMENTATION_NAME,
-            AnnotationSingletons::spanNameFromMethodRequest)
-        .addAttributesExtractor(
-            CodeAttributesExtractor.create(new MethodRequestCodeAttributesGetter()))
-        .addAttributesExtractor(
-            MethodSpanAttributesExtractor.create(
-                MethodRequest::method,
-                new WithSpanParameterAttributeNamesExtractor(),
-                MethodRequest::args))
-        .buildInstrumenter(AnnotationSingletons::spanKindFromMethodRequest);
-  }
-
-  private static SpanAttributesExtractor createAttributesExtractor() {
-    return SpanAttributesExtractor.create(new WithSpanParameterAttributeNamesExtractor());
-  }
-
-  private static SpanKind spanKindFromMethodRequest(MethodRequest request) {
-    return spanKindFromMethod(request.method());
-  }
-
-  private static SpanKind spanKindFromMethod(Method method) {
-    application.io.opentelemetry.instrumentation.annotations.WithSpan annotation =
-        method.getDeclaredAnnotation(
-            application.io.opentelemetry.instrumentation.annotations.WithSpan.class);
-    if (annotation == null) {
-      return SpanKind.INTERNAL;
-    }
-    return toAgentOrNull(annotation.kind());
-  }
-
-  private static SpanKind toAgentOrNull(
-      application.io.opentelemetry.api.trace.SpanKind applicationSpanKind) {
-    try {
-      return SpanKind.valueOf(applicationSpanKind.name());
-    } catch (IllegalArgumentException ignored) {
-      logger.log(FINE, "unexpected span kind: {0}", applicationSpanKind.name());
-      return SpanKind.INTERNAL;
-    }
-  }
-
-  private static String spanNameFromMethodRequest(MethodRequest request) {
-    return spanNameFromMethod(request.method());
-  }
-
-  private static String spanNameFromMethod(Method method) {
-    application.io.opentelemetry.instrumentation.annotations.WithSpan annotation =
-        method.getDeclaredAnnotation(
-            application.io.opentelemetry.instrumentation.annotations.WithSpan.class);
-    String spanName = annotation.value();
-    if (spanName.isEmpty()) {
-      spanName = SpanNames.fromMethod(method);
-    }
-    return spanName;
-  }
-
-  public static Context getContextForMethod(Method method) {
-    return inheritContextFromMethod(method) ? Context.current() : Context.root();
-  }
-
-  private static boolean inheritContextFromMethod(Method method) {
-    if (INHERIT_CONTEXT_METHOD_HANDLE == null) {
-      return true;
-    }
-
-    application.io.opentelemetry.instrumentation.annotations.WithSpan annotation =
-        method.getDeclaredAnnotation(
-            application.io.opentelemetry.instrumentation.annotations.WithSpan.class);
-    try {
-      return (boolean) INHERIT_CONTEXT_METHOD_HANDLE.invoke(annotation);
-    } catch (Throwable t) {
-      logger.log(FINE, "failed to read @WithSpan inheritContext()", t);
-      return true;
-    }
-  }
-
-  private AnnotationSingletons() {}
-}
