--- instrumentation/opentelemetry-instrumentation-annotations-1.16/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/instrumentationannotations/v1_16/MethodCodeAttributesGetter.java
@@ -1,22 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.instrumentationannotations.v1_16;
-
-import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesGetter;
-import java.lang.reflect.Method;
-
-final class MethodCodeAttributesGetter implements CodeAttributesGetter<Method> {
-
-  @Override
-  public Class<?> getCodeClass(Method method) {
-    return method.getDeclaringClass();
-  }
-
-  @Override
-  public String getMethodName(Method method) {
-    return method.getName();
-  }
-}
