--- instrumentation/opentelemetry-instrumentation-annotations-1.16/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/instrumentationannotations/v1_16/MethodRequest.java
@@ -1,26 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.instrumentationannotations.v1_16;
-
-import java.lang.reflect.Method;
-
-class MethodRequest {
-  private final Method method;
-  private final Object[] args;
-
-  MethodRequest(Method method, Object[] args) {
-    this.method = method;
-    this.args = args;
-  }
-
-  Method method() {
-    return this.method;
-  }
-
-  Object[] args() {
-    return this.args;
-  }
-}
