--- instrumentation/opentelemetry-instrumentation-annotations-1.16/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/instrumentationannotations/v1_16/MethodRequestCodeAttributesGetter.java
@@ -1,21 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.instrumentationannotations.v1_16;
-
-import io.opentelemetry.instrumentation.api.incubator.semconv.code.CodeAttributesGetter;
-
-final class MethodRequestCodeAttributesGetter implements CodeAttributesGetter<MethodRequest> {
-
-  @Override
-  public Class<?> getCodeClass(MethodRequest methodRequest) {
-    return methodRequest.method().getDeclaringClass();
-  }
-
-  @Override
-  public String getMethodName(MethodRequest methodRequest) {
-    return methodRequest.method().getName();
-  }
-}
