+++ instrumentation/opentelemetry-instrumentation-annotations-1.16/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/opentelemetry/instrumentation/annotations/v1_16/KotlinCoroutineUtil.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.opentelemetry.instrumentation.annotations.v1_16;
+
+import static net.bytebuddy.matcher.ElementMatchers.returns;
+
+import net.bytebuddy.description.method.MethodDescription;
+import net.bytebuddy.description.method.ParameterList;
+import net.bytebuddy.matcher.ElementMatcher;
+
+public class KotlinCoroutineUtil {
+
+  private KotlinCoroutineUtil() {}
+
+  public static ElementMatcher<MethodDescription> isKotlinSuspendMethod() {
+    // kotlin suspend methods return Object and take kotlin.coroutines.Continuation as last argument
+    return returns(Object.class)
+        .and(
+            target -> {
+              ParameterList<?> parameterList = target.getParameters();
+              if (!parameterList.isEmpty()) {
+                String lastParameter =
+                    parameterList.get(parameterList.size() - 1).getType().asErasure().getName();
+                return "kotlin.coroutines.Continuation".equals(lastParameter);
+              }
+              return false;
+            });
+  }
+}
